requirements.txt
torch>=2.2
torchvision>=0.17
gradio>=4.0
opencv-python
numpy
scikit-learn
Pillow